tasks/2_extract_opa_assessments_data/2.1_extract_opa_assessments_data/requirements.txt
functions_framework==3.5.0
protobuf==6.30.2
python-dotenv==1.1.0
Requests==2.32.3
